+++ src/ugdk/action/entity.h
@@ -6,7 +6,7 @@ namespace ugdk {
 class Entity {
   public:
     Entity() {}
-    virtual void Update(float dt) = 0;
+    virtual void Update(double dt) = 0;
 };
 
 }  // namespace ugdk
